CoverArt: Nullability, tweaks and cleanup.
Added JetBrains.Annotations and use [PublicAPI] and [UsedImplicitly] to replace the various [SuppressMessage]s.
Added a JsonUtils class.
Added a CoverArtEntity object, as base class that implements ICoverArtEntity; it "unwraps" the unhandled properties when applicable.
Cleaned up some of the XML Doc.
Enabled nullable reference types.
The FetchXXXImage[Async]() methods now never return null; an exception is thrown when no data is received (as was already the case for FetchXXXRelease[Async]()).
JSON deserialization now always goes via a string; this loses one of the speed advantages of System.Text.Json, but the payloads here are small enough for that not to matter too much. In addition, it avoids there being errors in a release build that cannot be reproduced in debug builds.