Switch to MetaBrainz.Build.Sdk
This drops the build submodule in favour of a NuGet SDK package
reference, and enables centralized package version management.

Conditional compilation has been updated to deal with the new set
of target frameworks this SDK brings.

It also bumps System.Text.Json to 5.0.0 and MetaBrainz.Common.Json
to 4.0.0 (for proper net5.0 support), with code updated accordingly.